pyproject.toml
[project]
name = "nonebot-plugin-alist"
version = "1.0.0"
description = "一个支持Alist多账号管理的NoneBot插件"
authors = [
    { name = "iam57", email = "me@iam57.com" },
]
dependencies = [
    "nonebot2>=2.3.2",
    "nonebot-adapter-onebot>=2.4.4",
    "httpx>=0.27.0",
    "tortoise-orm[aiosqlite]>=0.21.5",
    "aiosqlite>=0.17.0",
    "nonebot-plugin-alconna>=0.51.1",
]
requires-python = ">=3.10, <3.12"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = true
